Images now save as quality 80 jpegs. Gifs save straight from the InputStream. The csv read function had a bug. If an item has no description, nothing is saved.

+++ SimpleRss/src/main/java/com/poloure/simplerss/AdapterManageFilters.java
@@ -49,7 +49,7 @@ class AdapterManageFilters extends BaseAdapter
    public
    long getItemId(int position)
    {
-      return position;
+      return (long) position;
    }
 
    @Override
