+++ SimpleRss/src/main/java/com/poloure/simplerss/AdapterManageFragments.java
@@ -40,7 +40,7 @@ class AdapterManageFragments extends BaseAdapter
    public
    long getItemId(int position)
    {
-      return position;
+      return (long) position;
    }
 
    @Override
