+++ SimpleRss/src/main/java/com/poloure/simplerss/AdapterSettingsFunctions.java
@@ -50,7 +50,7 @@ class AdapterSettingsFunctions extends BaseAdapter
    public
    long getItemId(int position)
    {
-      return position;
+      return (long) position;
    }
 
    @Override
