+++ SimpleRss/src/main/java/com/poloure/simplerss/AdapterSettingsUi.java
@@ -48,7 +48,7 @@ class AdapterSettingsUi extends BaseAdapter
    public
    long getItemId(int position)
    {
-      return position;
+      return (long) position;
    }
 
    @Override
