+++ SimpleRss/src/main/java/com/poloure/simplerss/FeedItem.java
@@ -6,7 +6,7 @@ class FeedItem
    String m_itemUrl;
    String m_itemDescription;
    String m_imagePath;
-   int m_imageWidth;
-   int m_imageHeight;
+   short m_imageWidth;
+   short m_imageHeight;
    Long m_itemTime;
 }
